move things around2
