+++ admin-frontend/open_admin_app/lib/widgets/features/cell-view/value_cell.dart
@@ -2,9 +2,9 @@ import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:mrapi/api.dart';
+import 'package:open_admin_app/widgets/common/fh_tag.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/edit_feature_value_widget.dart';
 import 'package:open_admin_app/widgets/features/feature_dashboard_constants.dart';
-import 'package:open_admin_app/widgets/features/feature_value_status_tags.dart';
 import 'package:open_admin_app/widgets/features/per_application_features_bloc.dart';
 import 'package:open_admin_app/widgets/features/cell-view/flag_colored_on_off_label.dart';
 import 'package:open_admin_app/widgets/features/cell-view/strategy_tooltip.dart';
@@ -37,7 +37,8 @@ class ValueCellHolder extends StatelessWidget {
       );
     }
     if ((fv?.id == null) && efv.roles.isEmpty) {
-      return noAccessTag(null);
+      return const FHTagWidget(
+          text: 'NO ACCESS', state: TagStatus.disabled);
     }
     return const SizedBox.shrink();
   }
@@ -65,7 +66,6 @@ class _ValueContainer extends StatelessWidget {
             body: EditFeatureValueWidget(
                 fv: fv!,
                 environmentFeatureValue: efv,
-                featureValueType: feature.valueType!,
                 perApplicationFeaturesBloc: bloc,
                 feature: feature,
                 afv: afv,
