+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/boolean/edit_boolean_value_dropdown_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mrapi/api.dart';
-import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/custom_strategy_bloc.dart';
 class EditBooleanValueDropDownWidget extends StatefulWidget {
 
   const EditBooleanValueDropDownWidget({
