+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/json/edit_json_value_container.dart
@@ -6,7 +6,7 @@ import 'package:open_admin_app/widgets/common/fh_alert_dialog.dart';
 import 'package:open_admin_app/widgets/common/fh_flat_button.dart';
 import 'package:open_admin_app/widgets/common/fh_flat_button_transparent.dart';
 import 'package:open_admin_app/widgets/common/fh_json_editor.dart';
-import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/custom_strategy_bloc.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/json/json_viewer_field.dart';
 
 class EditJsonValueContainer extends StatefulWidget {
