+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/edit_value_strategy_link_button.dart
@@ -1,5 +1,5 @@
 import 'package:open_admin_app/widgets/common/fh_underline_button.dart';
-import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/per_feature_state_tracking_bloc.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/per_feature_state_tracking_bloc.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/individual_strategy_bloc.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/strategy_editing_widget.dart';
 import 'package:bloc_provider/bloc_provider.dart';
