+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/percentage_utils.dart
