+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/strategy_card_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mrapi/api.dart';
-import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/feature_dashboard_constants.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/feature_dashboard_constants.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/delete_strategy_icon_button.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/edit_value_strategy_link_button.dart';
