+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature-value/strategies/strategy_editing_widget.dart
@@ -7,7 +7,7 @@ import 'package:open_admin_app/widgets/common/fh_flat_button.dart';
 import 'package:open_admin_app/widgets/common/fh_flat_button_transparent.dart';
 import 'package:open_admin_app/widgets/common/fh_outline_button.dart';
 import 'package:open_admin_app/widgets/common/input_fields_validators/input_field_number_formatter.dart';
-import 'package:open_admin_app/widgets/features/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/percentage_utils.dart';
+import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/custom_strategy_bloc.dart';import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/percentage_utils.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/individual_strategy_bloc.dart';
 import 'package:open_admin_app/widgets/features/edit-feature-value/strategies/strategy_utils.dart';
