+++ admin-frontend/open_admin_app/lib/widgets/features/edit-feature/delete_feature_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:mrapi/api.dart';
 import 'package:openapi_dart_common/openapi.dart';
 
-import 'per_application_features_bloc.dart';
+import '../per_application_features_bloc.dart';
 
 class FeatureDeleteDialogWidget extends StatelessWidget {
   final Feature feature;
