+++ admin-frontend/open_admin_app/lib/widgets/features/feature_dashboard_constants.dart
@@ -1,12 +1,5 @@
 import 'package:flutter/material.dart';
 
-const unselectedRowHeight = 70.0;
-const unselectedRowHeightPerStrategy = 45.0;
-const lockHeight = 60.0;
-const selectedRowHeight = 260.0;
-const selectedRowHeightPerStrategy = 50.0;
-const headerHeight = 40.0;
-const cellWidth = 250.0;
 const defaultValueColor = Color(0xffE0EEFD);
 const defaultValueColorDark = Color(0xff303030);
 const defaultTextColor = Color(0xff4DA1FF);
@@ -14,3 +7,7 @@ const strategyValueColor = Color(0xffE6F7EE);
 const strategyValueColorDark = Color(0xff545669);
 const strategyTextColor = Color(0xff2DCD7A);
 const rowsPerPage = 5;
+const featureNameBoxWidth = 150.0;
+const featureValueCellWidth = 150.0;
+const featureNameCellWidth = 200.0;
+const tableHeight = 600.0;
